Move errors pkg into backend
