+++ backend/cmd/backend/main.go
@@ -10,7 +10,7 @@ import (
 	texporter "github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/trace"
 	"github.com/nownabe/golink/backend"
 	"github.com/nownabe/golink/backend/clog"
-	"github.com/nownabe/golink/go/errors"
+	"github.com/nownabe/golink/backend/errors"
 	"go.opentelemetry.io/contrib/detectors/gcp"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/propagation"
