+++ backend/golink_service.go
@@ -9,9 +9,9 @@ import (
 	"cloud.google.com/go/firestore"
 	"github.com/bufbuild/connect-go"
 	"github.com/nownabe/golink/backend/clog"
+	"github.com/nownabe/golink/backend/errors"
 	golinkv1 "github.com/nownabe/golink/backend/gen/golink/v1"
 	"github.com/nownabe/golink/backend/golinkcontext"
-	"github.com/nownabe/golink/go/errors"
 	"golang.org/x/exp/slices"
 )
