+++ backend/interceptor/interceptors.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/bufbuild/connect-go"
 	"github.com/nownabe/golink/backend/clog"
-	"github.com/nownabe/golink/go/errors"
+	"github.com/nownabe/golink/backend/errors"
 )
 
 const (
