+++ backend/middleware/request_id.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/nownabe/golink/backend/clog"
 	"github.com/nownabe/golink/backend/clog/clogcontext"
-	"github.com/nownabe/golink/go/errors"
+	"github.com/nownabe/golink/backend/errors"
 	"go.opentelemetry.io/otel/trace"
 )
