+++ backend/redirect_handler.go
@@ -10,7 +10,7 @@ import (
 
 	"cloud.google.com/go/firestore"
 	"github.com/nownabe/golink/backend/clog"
-	"github.com/nownabe/golink/go/errors"
+	"github.com/nownabe/golink/backend/errors"
 )
 
 type redirectHandler struct {
